% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ABandC}
\alias{ABandC}
\title{Takes a string of words and combines them into a sentance that lists them.}
\usage{
ABandC(x)
}
\arguments{
\item{x}{Character strings you want in your string.}
}
\description{
This function alows you to take a string of words and combine them into a sentance list. For example, 'apples', 'oranges', 'pears' would become 'apples, oranges, and pears'. This function uses oxford commas.
}
\examples{
ABandC(c(1,2,"hello",4,"world",6))
}
\keyword{list,}
\keyword{strings}
