% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addfootnotes}
\alias{addfootnotes}
\title{Add footnotes to table.}
\usage{
addfootnotes(
  df.dat,
  table0 = NA,
  state0 = NA,
  region0 = NA,
  species0 = NA,
  footnote0
)
}
\arguments{
\item{df.dat}{The data frame you want to add the footnote to.}

\item{table0}{Name of the table you want the footnote added to. Default: NA.}

\item{state0}{The state you want the footnote attached to. Default: NA.}

\item{region0}{The region you want the footnote attached to. Default: NA.}

\item{species0}{The species you want the footnote attached to. Default: NA.}

\item{footnote0}{The list of footnotes you want to be added to.}
}
\description{
This function adds a footnote to a dataframe.
}
\examples{
addfootnotes()
}
\keyword{footnote,}
\keyword{footnotes}
\keyword{matrix,}
