% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{ageoffile}
\alias{ageoffile}
\title{Find the age of the file, when it was created.}
\usage{
ageoffile(path, format = "\%B \%d, \%Y")
}
\arguments{
\item{path}{Path to the file.}

\item{format}{default = "\%B \%d, \%Y"}
}
\description{
Find the age of the file, when it was created.
}
\keyword{Data}
\keyword{Management}
