% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{auto_counter}
\alias{auto_counter}
\title{Add a counter number.}
\usage{
auto_counter(counter0)
}
\arguments{
\item{counter0}{The value it was to be added 1 to}
}
\description{
Add a counter number, 1, 1+1=2, 2+1=3.
}
\examples{
auto_counter()
}
\keyword{Data}
\keyword{Management}
