% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{buildTM}
\alias{buildTM}
\title{Build your intitial architecture for your new NOAA Tech Memo or Report}
\usage{
buildTM(
  sections = c("frontmatter", "abstract", "introduction", "methods", "results",
    "discussion", "workscited", "workscitedR"),
  support_scripts = c("directories", "functions", "dataDL", "data"),
  authors = "",
  title = "",
  pptxstylesreference.pptx = "refdoc_NMFS_ppt_template",
  wordstylesreference.docx = "refdoc_NOAATechMemo.docx",
  csl = "apa"
)
}
\arguments{
\item{sections}{a string of the different sections of your TM. Default = c("frontmatter", "abstract", "introduction", "methods", "results", "discussion", "workscited", "workscitedR"). Note that frontmatter and workcitedR both have specific templates, and all others are from a blank template.}

\item{support_scripts}{To make sure we nice and neatly compartemantalize our work, create the below supporting .R files that you will source into your 'run' file. Default = c("functions", "dataDL", "data")}

\item{authors}{Default = "". Here, add your First Lastname (email).}

\item{title}{Default = "". Here, put the title of your report.}

\item{pptxstylesreference.pptx}{Default = "refdoc_NMFS_ppt_template". Alternatively, you can add the path here to your own reference pptx. You can also opt out of a powerpoint by setting this parameter to NA.}

\item{wordstylesreference.docx}{Document style reference guide is essentially a word document where you have defined each style. Either use a local document (insert "path") or some of the pre-made templates ("NOAATechMemo" or "FisheriesEconomicsOfTheUS"). Default = "NOAATechMemo"}

\item{csl}{Citation style. Either use a local document (insert "path") or some of the pre-made templates ("apa", "new-phytologist"). A NOAA TM citation style needs to be created, but until then, the default = "apa".}
}
\description{
Build your intitial architecture for your new NOAA Tech Memo or Report
}
\examples{
# not run:
# buildTM()
}
