% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{create_metadata}
\alias{create_metadata}
\title{Record Metadata}
\usage{
create_metadata(dir_out = ".", title = "My Project")
}
\arguments{
\item{dir_out}{Path file will be saved to.}

\item{title}{Title of file.}
}
\description{
Record Metadata
}
\keyword{metadata}
