% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CreateLoadedDataMetadata}
\alias{CreateLoadedDataMetadata}
\title{Create Metadata.}
\usage{
CreateLoadedDataMetadata(dir.out, data, title)
}
\arguments{
\item{dir.out}{Path file will be saved to.}

\item{data}{The data.frame.}

\item{title}{Title of file.}
}
\description{
Create Metadata.
}
\examples{
CreateLoadedDataMetadata()
}
\keyword{metadata}
