% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{CreateMetadata}
\alias{CreateMetadata}
\title{Create CreateMetadata}
\usage{
CreateMetadata(dir.out = ".", title = "My Project")
}
\arguments{
\item{dir.out}{Path file will be saved to.}

\item{title}{Title of file.}
}
\description{
Create Metadata.
}
\keyword{metadata}
