% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{df2js}
\alias{df2js}
\title{Convert dataframe to javascript}
\usage{
df2js(dat)
}
\arguments{
\item{dat}{The data frame you want to add the footnote to.}
}
\description{
Convert dataframe to javascript matrix.
}
\examples{
dat <- cbind.data.frame(matrix(LETTERS[1:8], nrow = 4), 
                           matrix(1:8, nrow = 4))
df2js(dat = dat)
}
\keyword{data.frame,}
\keyword{footnote}
\keyword{footnotes,}
\keyword{javascript,}
