% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{df2list}
\alias{df2list}
\title{Convert data.frame to list.}
\usage{
df2list(dat, listlevels, collect)
}
\arguments{
\item{dat}{A data.frame.}

\item{listlevels}{How many levels should be in your list.}

\item{collect}{Something.}
}
\description{
Convert data.frame to list.
}
\examples{
df2list()
}
\keyword{Data}
\keyword{Management}
