% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FilenameAuto}
\alias{FilenameAuto}
\title{Auto-formating names of files.}
\usage{
FilenameAuto(xreg, xsect, xstate, xdesc)
}
\arguments{
\item{xreg}{region number.}

\item{xsect}{state number.}

\item{xstate}{state number.}

\item{xdesc}{Description.}
}
\description{
Auto-formating names of files.
}
\examples{
FilenameAuto()
}
\keyword{file}
\keyword{saving}
