% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{firstup}
\alias{firstup}
\title{Make first letter of a string capitalized.}
\usage{
firstup(x)
}
\arguments{
\item{x}{The text string.}
}
\description{
Make first letter of a string capitalized.
}
\examples{
firstup()
}
\keyword{Text}
\keyword{editing}
