% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_cells}
\alias{format_cells}
\title{Add bold, italics, strikethrough in formating to table.}
\usage{
format_cells(df.dat, rows, cols, fonttype)
}
\arguments{
\item{df.dat}{A data.frame.}

\item{rows}{The rows you want to apply formatting to.}

\item{cols}{The columns you want to apply formatting to.}

\item{fonttype}{fonttype = c("italics", "bold", "strikethrough").}
}
\description{
Add bold, italics, strikethrough in formating to table.
}
\examples{
format_cells()
}
\keyword{Modify}
\keyword{number,}
\keyword{units}
