% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{funct_df2js}
\alias{funct_df2js}
\title{Convert dataframe to javascript}
\usage{
funct_df2js(df.dat)
}
\arguments{
\item{df.dat}{The data frame you want to add the footnote to.}
}
\description{
Convert dataframe to javascript matrix.
}
\examples{
df.dat <- cbind.data.frame(matrix(LETTERS[1:8], nrow = 4), 
                           matrix(1:8, nrow = 4))
funct_df2js(df.dat = df.dat)
}
\keyword{data.frame,}
\keyword{footnote}
\keyword{footnotes,}
\keyword{javascript,}
