% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{funct_save}
\alias{funct_save}
\title{Save files in several different ways.}
\usage{
funct_save(
  Tfootnotes = NA,
  place,
  area,
  folder,
  maxyr,
  minyr,
  csvname1000,
  csvname,
  st,
  reg,
  xreg,
  xsect,
  xstate,
  dir.outputtables,
  statereg,
  temp.code = NA,
  temp.print = NA,
  webtool = NA
)
}
\arguments{
\item{Tfootnotes}{Default = NA.}

\item{place}{region}

\item{area}{state}

\item{folder}{path}

\item{maxyr}{Maxium year being assessed.}

\item{minyr}{Minimum year being assessed.}

\item{csvname1000}{Title name for values in thousands.}

\item{csvname}{Name of the CSV.}

\item{st}{state}

\item{reg}{region}

\item{xreg}{region number}

\item{dir.outputtables}{region}

\item{statereg}{data.frame of files}

\item{temp.code}{data.frame of basic table.Default = NA.}

\item{temp.print}{data.frame of table made for print, all fancy like. Default = NA.}

\item{webtool}{data.frame of table made for webtool. Default = NA.}

\item{ending}{File Ending.}

\item{xst}{state number}
}
\description{
Save files in several different ways.
}
\examples{
funct_save()
}
\keyword{file}
\keyword{saving}
