% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{googledrive_txt_dl}
\alias{googledrive_txt_dl}
\title{Download text from google drive as plain text}
\usage{
googledrive_txt_dl(
  filename_gd,
  filename_dl = "googledrive_dl_text",
  path = "./",
  verbose = TRUE
)
}
\arguments{
\item{filename_gd}{The file path or filename of the google doc from google drive.}

\item{filename_dl}{The filename you want to download the google doc as. Default = "googledrive_dl_text".}

\item{path}{The path you want to save this file to. Default = "./"}

\item{verbose}{Logical, indicating whether to print informative messages (default TRUE).}
}
\value{
The plain text from the google doc.
}
\description{
Download text from google drive as plain text
}
\examples{
googledrive::drive_auth() # Using the first token you have.
1

txt <- googledrive_txt_dl(filename_gd = "test123123_doc",
                          filename_dl = "test123123_dl",
                          verbose = FALSE)
txt
# for good file keeping, I'll delete these
file.remove('test123123_dl.txt', 'test123123_dl.zip')
}
