% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{is_something_in_this_matrix}
\alias{is_something_in_this_matrix}
\title{Is something in a matrix? Let's check!}
\usage{
is_something_in_this_matrix(x, search_for)
}
\arguments{
\item{x}{The matrix that needs to be searched.}

\item{search_for}{Items to be searched for in matrix x.}
}
\value{
TRUE or FALSE
}
\description{
This function searches to see if item 'search_for' is within the matrix 'x' and returns a respective TRUE (T) and FALSE (F).
}
\examples{
is_something_in_this_matrix(x = data.frame(matrix(1:9, nrow = 3, ncol = 3)), 
                       search_for = 9)
is_something_in_this_matrix(x = data.frame(matrix(LETTERS[1:9], nrow = 3, ncol = 3)), 
                       search_for = "J")
}
\keyword{footnote,}
\keyword{footnotes}
\keyword{matrix,}
\keyword{search,}
