% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list2string_ft}
\alias{list2string_ft}
\title{Convert list to string for footnotes}
\usage{
list2string_ft(x)
}
\arguments{
\item{x}{The data frame you want to add the footnote to.}
}
\description{
This function adds a footnote to a dataframe.
}
\examples{
list2string_ft()
}
\keyword{footnote,}
\keyword{footnotes}
