% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list2string_webtool}
\alias{list2string_webtool}
\title{Convert list to string for webtool.}
\usage{
list2string_webtool(x)
}
\arguments{
\item{x}{The data frame you want to add the footnote to.}
}
\description{
This function adds a footnote to a dataframe.
}
\examples{
list2string_webtool()
}
\keyword{footnote,}
\keyword{footnotes}
\keyword{webtool,}
