% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{mod_number}
\alias{mod_number}
\title{Modify numbers.}
\usage{
mod_number(x, divideby = 1000, comma_seperator = TRUE, digits = 0)
}
\arguments{
\item{x}{A numeric.}

\item{divideby}{The value you want all of your values divided by. Default = 1000.}

\item{comma_seperator}{Do you want numbers to have commas in it ("1,000" (T) vs. "1000" (F). Default = T.}

\item{digits}{How many digits you would like your number to have. Default = 0.}
}
\description{
Modify numbers.
}
\examples{
x = data.frame(matrix(data = c(20000678660, 234567, 1, NA, 2345, 23),
       ncol = 2))
mod_number(x)
mod_number(x,
       comma_seperator = FALSE)
x = data.frame(matrix(data = c(200000, 234567, 1, NA, 2345, 23)))
mod_number(x,
       divideby = 1,
       digits = 2)
}
\keyword{Modify}
\keyword{number}
