% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modnum}
\alias{modnum}
\title{Modify number.}
\usage{
modnum(x, divideby = 1000, commaseperator = T, digits = 0)
}
\arguments{
\item{x}{A numeric.}

\item{divideby}{The value you want all of your values divided by. Default = 1000.}

\item{commaseperator}{Do you want numbers to have commas in it ("1,000" (T) vs. "1000" (F). Default = T.}

\item{digits}{How many digits you would like your number to have. Default = 0.}
}
\description{
Modify number.
}
\examples{
modnum()
}
\keyword{Modify}
\keyword{number}
