% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{numbers2words}
\alias{numbers2words}
\title{Convert number to text string.}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{The numbers that need to be converted to string.}
}
\description{
Function by John Fox found here: http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html and https://github.com/ateucher/useful_code/blob/master/R/numbers2words.r
}
\examples{
numbers2words(x = 1890)
numbers2words(x = 3)
numbers2words(x = 1800090)
}
\keyword{Text}
\keyword{editing}
