% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{numbers2words_th}
\alias{numbers2words_th}
\title{Convert number to text string.}
\usage{
numbers2words_th(x, type = "word")
}
\arguments{
\item{x}{The numbers that need to be converted to string.}

\item{type}{How the numbers should be converted. Default = "word" (which produces "fifty-third"), but you can also select "val_th" (which produces "53rd").}
}
\description{
Convert number to text string to the 'st, 'nd, 'rd, or 'th.
}
\examples{
numbers2words_th(x = 3, type = "val_th")
numbers2words_th(x = 3, type = "word")
}
\keyword{Text}
\keyword{editing}
