% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{pchange}
\alias{pchange}
\title{Calculate the percent change.}
\usage{
pchange(start, end, ending = "", percent_first = TRUE)
}
\arguments{
\item{start}{The value it started with.}

\item{end}{The value it ended with.}

\item{ending}{A text string. Default "".}

\item{percent_first}{Options: T/F.}
}
\description{
Calculate the percent change.
}
\examples{
pchange(start = 8, end = 1)
pchange(start = 3, end = 6, ending = " in fish landings", percent_first = TRUE)
pchange(start = 3, end = 4, ending = " in fish landings", percent_first = FALSE)
}
\keyword{Modify}
\keyword{number}
