% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{save_equations}
\alias{save_equations}
\title{Systematically save your figures for your report}
\usage{
save_equations(
  equation,
  list_equations,
  header = "",
  footnote = "",
  cnt_chapt_content = "001",
  cnt = 1,
  type = "Equation",
  alttext = "",
  nickname = "",
  message = FALSE
)
}
\arguments{
\item{equation}{The latex equation you would like to be saved.}

\item{list_equations}{The list where all equations will be saved.}

\item{header}{The name and title of the figure. Default = "".}

\item{footnote}{Any footnote you want attached to this figure.}

\item{cnt_chapt_content}{The order number that this exists in the chapter.}

\item{cnt}{The figure number.}

\item{type}{Default = "Equation", but can be anything that the element needs to be called (e.g., "Eq.", "Equ.") to fit in the phrase "Equation 1. This is my equation!".}

\item{alttext}{String with what the alternative text is.}

\item{nickname}{A unique name that can be used to identify the figure so it can be referenced later in the report.}

\item{message}{TRUE/FALSE. Default = FALSE. If TRUE, it will print information about where your plot has been saved to.}
}
\value{
list_equations  updated with the new equation and metadata.
}
\description{
Systematically save your figures for your report
}
\examples{
list_equations  <- c()
cnt_eq <- 0

cnt_eq<-NMFSReports::auto_counter(cnt_eq)
list_equations <-NMFSReports::save_equations(
   equation = "$$c^2 = b^2 + a^2$$",
   cnt = cnt_eq,
   list_equations  = list_equations ,
   nickname = "pythagorean",
   header = "Pythagorean theorem",
   footnote = "footnote about how cool the pythagorean theorem is.",
   alttext = "The Pythagoras theorem is a mathematical law that states that the sum of squares of the lengths of the two short sides of the right triangle is equal to the square of the length of the hypotenuse.")

cnt_eq<-NMFSReports::auto_counter(cnt_eq)
list_equations <-NMFSReports::save_equations(
   equation = "$$F = G \frac{m_1 m_2}{d^2}$$",
   cnt = cnt_eq,
   list_equations  = list_equations ,
   nickname = "Newton",
   header = "Newton's Universal Law of Gravitation")

names(list_equations )
list_equations
}
