% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{save_tables}
\alias{save_tables}
\title{Systematically save your report tables for your report}
\usage{
save_tables(
  table_raw = NULL,
  table_print = NULL,
  table_list = c(),
  header = "",
  footnote = "",
  filename0 = "x",
  cnt_chapt_content = "001",
  cnt = 1,
  path = NULL,
  output_type = c("csv"),
  type = "Table",
  filename_desc = ""
)
}
\arguments{
\item{table_raw}{Optional. The data.frame that has no rounding and no dividing of numbers (good to save this for record keeping). Default = NA.}

\item{table_print}{The data.frame as table will be seen in the report.}

\item{table_list}{Save tables in a list}

\item{header}{The name and title of the figure. Default = "".}

\item{footnote}{Any footnote you want attached to this figure.}

\item{filename0}{The filename set at the begining of the chapter}

\item{cnt_chapt_content}{The order number that this exists in the chapter.}

\item{cnt}{The figure number}

\item{path}{The path the file needs to be saved to. Default = "NULL", meaning it wont save anything and will override all other saving elements.}

\item{output_type}{Default = c("csv"). Can be anything supported by utils::write.table.}

\item{type}{Default = "Table", but can be anything that the element needs to be called (e.g., "Graphic", "Fig.", "Graph") to fit in the phrase "Table 1. This is my spreadsheet!". Always save in pdf so you can make last minute edits in adobe acrobat!}

\item{filename_desc}{Additional description text for the filename that will be added at the name of file before the filename extention, before the "_raw" or "_print". Default = "". Can be use to add a species name, location, or anything else that would make it easier to know what that file shows.}
}
\description{
Systematically save your report tables for your report
}
\examples{
# Select data and make plot
table_raw<-data.frame(x = rnorm(n = 10), 
                      y = rnorm(n = 10), 
                      col = rep_len(x = c("a", "b"), length.out = 5)) 
table_print <- table_raw
table_print[,c("x", "y")] <- NMFSReports::mod_number(table_print[,c("x", "y")], 
                                                     divideby = 1, 
                                                     comma_seperator = TRUE, 
                                                     digits = 2)
save_tables(table_raw = table_raw, 
           table_print=table_print, 
           header = "Here is a table!", 
           footnote = "A footnote for this table!")
}
