% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simpleCap}
\alias{simpleCap}
\title{Make a string sentance case.}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{The text string.}
}
\description{
Make a string sentance case.
}
\examples{
simpleCap()
}
\keyword{Text}
\keyword{editing}
