% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{stndth}
\alias{stndth}
\title{Find the 'st, 'nd, or 'th of a value}
\usage{
stndth(x)
}
\arguments{
\item{x}{a value you want the 'st, 'nd, or 'th of}
}
\value{
a character string of the appropriate 'st, 'nd, or 'th
}
\description{
Find the 'st, 'nd, or 'th of a value
}
\examples{
stndth(3)
stndth(11)
stndth(112)
stndth(1120)
}
