% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{temp_save}
\alias{temp_save}
\title{Save files in several different ways.}
\usage{
temp_save(
  df.dat,
  minyr,
  maxyr,
  filename00,
  ending,
  csvname,
  Tfootnotes = NA,
  reg,
  st
)
}
\arguments{
\item{df.dat}{The data frame you want to add the footnote to.}

\item{minyr}{Minimum year being assessed.}

\item{maxyr}{Maxium year being assessed.}

\item{filename00}{filename00.}

\item{ending}{File Ending.}

\item{csvname}{Name of the CSV.}

\item{Tfootnotes}{Default = NA.}

\item{reg}{region}

\item{st}{state}
}
\description{
Save files in several different ways.
}
\examples{
temp_save()
}
\keyword{file}
\keyword{saving}
