% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{text_increasingdecreasing}
\alias{text_increasingdecreasing}
\title{List stocks and calculate their absolute or nominal increase or decrease.}
\usage{
text_increasingdecreasing(
  minyr,
  maxyr,
  dat,
  metric,
  orgby = "pct",
  absvalue = F,
  pctvalue = T,
  preunit = "",
  postunit = "",
  decreasingTF,
  section,
  capitalizefirst = F,
  foot = "",
  bullets
)
}
\arguments{
\item{minyr}{The data frame you want to add the footnote to.}

\item{maxyr}{Minimum year being assessed.}

\item{dat}{Maxium year being assessed.}

\item{metric}{Maxium year being assessed.}

\item{orgby}{Maxium year being assessed.}

\item{absvalue}{The year being assessed.}

\item{pctvalue}{Maxium year being assessed. Options equal integers, such that 1 = first, 2 = ...}

\item{preunit}{String of text for text that should go after the value. E.g., 'pre ##\%  decrease'. Default = "".}

\item{postunit}{String of text for text that should go after the value. E.g., '##\% post decrease'. Default = "".}

\item{decreasingTF}{Should this be in decreasing (T) or increasing (F) order? Options: T, F.}

\item{section}{Default = 'Nominal'.}

\item{capitalizefirst}{Default = TRUE.}

\item{foot}{Default = "".}

\item{bullets}{Default = TRUE.}
}
\description{
List stocks and calculate their absolute or nominal increase or decrease.
}
\examples{
text_increasingdecreasing()
}
\keyword{Text}
\keyword{data.frame,}
\keyword{editing}
\keyword{javascript,}
