% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{text_increasingdecreasingbullets}
\alias{text_increasingdecreasingbullets}
\title{List stocks and calculate their absolute or nominal increase or decrease in bullet form.}
\usage{
text_increasingdecreasingbullets(
  minyr,
  maxyr,
  dat,
  orgby = orgby,
  absvalue,
  pctvalue,
  preunit = preunit,
  postunit = postunit,
  decreasingTF,
  section = section,
  capitalizefirst = T,
  keyphrase
)
}
\arguments{
\item{minyr}{The data frame you want to add the footnote to.}

\item{maxyr}{Minimum year being assessed.}

\item{dat}{Maxium year being assessed.}

\item{orgby}{Maxium year being assessed.}

\item{absvalue}{The year being assessed.}

\item{pctvalue}{Maxium year being assessed. Options equal integers, such that 1 = first, 2 = ...}

\item{preunit}{String of text for text that should go after the value. E.g., 'pre ##\%  decrease'. Default = "".}

\item{postunit}{String of text for text that should go after the value. E.g., '##\% post decrease'. Default = "".}

\item{decreasingTF}{Should this be in decreasing (T) or increasing (F) order? Options: T, F.}

\item{section}{Default = 'Nominal'.}

\item{capitalizefirst}{Default = TRUE.}

\item{keyphrase}{.}
}
\description{
List stocks and calculate their absolute or nominal increase or decrease in bullet form.
}
\examples{
text_increasingdecreasingbullets()
}
\keyword{Text}
\keyword{data.frame,}
\keyword{editing}
