% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{text_liststocks}
\alias{text_liststocks}
\title{List stocks and calculate their absolute or nominal increase or decrease.}
\usage{
text_liststocks(
  dat,
  metric,
  orgby = "pct",
  absvalue = F,
  pctvalue = T,
  preunit = "",
  postunit = "",
  decreasingTF,
  yr,
  lv,
  section = "nominal",
  capitalizefirst = T
)
}
\arguments{
\item{dat}{The data frame you want to add the footnote to.}

\item{metric}{Minimum year being assessed.}

\item{orgby}{Maxium year being assessed.}

\item{absvalue}{Maxium year being assessed.}

\item{pctvalue}{Maxium year being assessed.}

\item{preunit}{String of text for text that should go after the value. E.g., 'pre ##\%  decrease'.}

\item{postunit}{String of text for text that should go after the value. E.g., '##\% post decrease'.}

\item{decreasingTF}{Should this be in decreasing (T) or increasing (F) order? Options: T, F.}

\item{yr}{The year being assessed.}

\item{lv}{Maxium year being assessed. Options equal integers, such that 1 = first, 2 = ...}

\item{section}{Default = 'Nominal'.}

\item{capitalizefirst}{Default = TRUE.}
}
\description{
List stocks and calculate their absolute or nominal increase or decrease.
}
\examples{
text_liststocks()
}
\keyword{Text}
\keyword{data.frame,}
\keyword{editing}
\keyword{javascript,}
