% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TitleCase}
\alias{TitleCase}
\title{Make a String Title Case}
\usage{
TitleCase(str)
}
\arguments{
\item{str}{A string that you want to be in title case}
}
\description{
Make a String Title Case (making and, the, an, etc. lower case)
}
\examples{
TitleCase()
}
\keyword{Case,}
\keyword{Title,}
\keyword{strings}
\keyword{word}
