% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tolower2}
\alias{tolower2}
\title{Make a string lower case except for stated proper nouns.}
\usage{
tolower2(str0, capitalizefirst = F)
}
\arguments{
\item{str0}{The text string.}

\item{capitalizefirst}{Default = FALSE}
}
\description{
Make a string lower case except for stated proper nouns.
}
\examples{
tolower2()
}
\keyword{Text}
\keyword{editing}
