% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ageoffile}
\alias{ageoffile}
\title{Find the age of the file, when it was created.}
\usage{
ageoffile(path)
}
\arguments{
\item{path}{Path to the file.}
}
\description{
Find the age of the file, when it was created.
}
\examples{
ageoffile()
}
\keyword{Data}
\keyword{Management}
