% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{buildTM}
\alias{buildTM}
\title{Build your intitial architecture for your new NOAA Tech Memo or Report}
\usage{
buildTM(
  sections = c("frontmatter", "abstract", "introduction", "methods", "results",
    "discussion", "workscited", "workscitedR", "presentation"),
  support_scripts = c("directories", "functions", "dataDL", "data"),
  authors = "",
  title = "",
  styles_reference_pptx = "refppt_nmfs",
  styles_reference_docx = "refdoc_noaa_tech_memo",
  bibliography.bib = "bib_example",
  csl = "apa"
)
}
\arguments{
\item{sections}{a string of the different sections of your TM. Default = c("frontmatter", "abstract", "introduction", "methods", "results", "discussion", "workscited", "endmatter"). Note that "frontmatter" and "endmatter both have specific templates, and all others are from a blank template. "endmatter" will document all of your citations throughout the report, the R packages you used to create this report. I'm biased, but please give credit where credit is due! There are also spots here to list people's O}

\item{support_scripts}{To make sure we nice and neatly compartemantalize our work, create the below supporting .R files that you will source into your 'run' file. Default = c("functions", "dataDL", "data")}

\item{authors}{Default = "". Here, add your First Lastname.}

\item{title}{Default = "". Here, put the title of your report.}

\item{styles_reference_pptx}{A style reference guide from a powerpoint document (.pptx). This pulls the styles from a powerpoint document where you have defined each style. Either use a local document (insert full "path") or some of the pre-made templates ("refppt_nmfs"). Default = "nmfs".}

\item{styles_reference_docx}{A style reference guide from a word document (.docx). This pulls the styles from a word document where you have defined each style. Either use a local document (insert full "path") or some of the pre-made templates ("refdoc_noaa_tech_memo" or "refdoc_fisheries_economics_of_the_us"). Default = "refdoc_noaa_tech_memo".}

\item{bibliography.bib}{Either use a local document (.bib format; insert full "path") or the example file from the package ("bib_example"). Default = "bib_example".}

\item{csl}{Citation style. Either use a local document (insert "path") or some of the pre-made templates ("apa", "new-phytologist"). A NOAA TM citation style needs to be created, but until then, the default = "apa".}
}
\description{
Build your intitial architecture for your new NOAA Tech Memo or Report
}
\examples{
sections = c("frontmatter", "abstract", "introduction", "methods", "results",
            "discussion", "workscited", "workscitedR", "presentation")
support_scripts = c("directories", "functions", "dataDL", "data")
authors = "Me, Myself, and I"
title = "Awesome Report!"
styles_reference_pptx = "refppt_nmfs"
styles_reference_docx = "refdoc_noaa_tech_memo"
bibliography.bib = "bib_example"
csl = "apa"

# not run:
# buildTM(
#   sections = sections,
#   support_scripts = support_scripts,
#   authors = authors,
#   title = title,
#   styles_reference_pptx = styles_reference_pptx,
#   styles_reference_docx = styles_reference_docx,
#   bibliography.bib = bibliography.bib,
#   csl = csl
# )
}
