% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CreateMetadata}
\alias{CreateMetadata}
\title{Create CreateMetadata}
\usage{
CreateMetadata(dir.out, title)
}
\arguments{
\item{dir.out}{Path file will be saved to.}

\item{title}{Title of file.}
}
\description{
Create Metadata.
}
\examples{
CreateMetadata()
}
\keyword{metadata}
