% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{crossref}
\alias{crossref}
\title{Reference a figure or table}
\usage{
crossref(list_obj, nickname, sublist = "number")
}
\arguments{
\item{list_obj}{A list object created by list_figures or list_tables.}

\item{nickname}{A unique string that is used to identify the plot or table in list_figures or list_tables, respectively.}

\item{sublist}{A string of the sublist in list_figures or list_tables you want the contents returned from.}
}
\value{
The item in the list.
}
\description{
Reference a figure or table
}
\examples{
list_figures <- c()
pp <- plot(x = 1, y = 1)
list_figures <- NMFSReports::save_figures(
   figure = pp,
   list_figures = list_figures,
   header = "blah blah blah",
   nickname = "example_figure", # a unique name you can refer back to
   cnt_chapt_content = "003",
   cnt = "012")
list_figures <- NMFSReports::save_figures(
   figure = pp,
   list_figures = list_figures,
   header = "blah blah blah",
   nickname = "example_figure_other", # a unique name you can refer back to
   cnt_chapt_content = "003",
   cnt = "013")
list_figures
refnum <- NMFSReports::crossref(
   list_obj = list_figures,
   nickname = "example_figure",
   sublist = "number")
refnum
print(paste0("Please refer to figure ", refnum,
             " to see this figure, not the other figure."))
}
