% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{format_cells}
\alias{format_cells}
\title{Add bold, italics, strikethrough in formating to table.}
\usage{
format_cells(dat, rows, cols, fonttype)
}
\arguments{
\item{dat}{A data.frame.}

\item{rows}{The rows you want to apply formatting to.}

\item{cols}{The columns you want to apply formatting to.}

\item{fonttype}{fonttype = c("italics", "bold", "strikethrough").}
}
\description{
https://stackoverflow.com/questions/28166168/how-to-change-fontface-bold-italics-for-a-cell-in-a-kable-table-in-rmarkdown
}
\examples{
df <- data.frame(char = c('a','b','c'),
                 num = c(1,2,3))

format_cells(df, 1, 1, "italics")
format_cells(df, 2, 2, "bold")
format_cells(df, 3, 1:2, "strikethrough")

library(knitr)
library(kableExtra)
library(magrittr)
df \%>\%
  format_cells(1, 1, "italics") \%>\%
  format_cells(2, 2, "bold") \%>\%
  format_cells(3, 1:2, "strikethrough") \%>\%
  knitr::kable()
}
\keyword{Modify}
\keyword{number,}
\keyword{units}
