% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{funct_df2js}
\alias{funct_df2js}
\title{Convert dataframe to javascript}
\usage{
funct_df2js(df.dat, minyr, maxyr)
}
\arguments{
\item{df.dat}{The data frame you want to add the footnote to.}

\item{minyr}{Minimum year being assessed.}

\item{maxyr}{Maxium year being assessed.}
}
\description{
Convert dataframe to javascript matrix.
}
\examples{
funct_df2js()
}
\keyword{data.frame,}
\keyword{footnote}
\keyword{footnotes,}
\keyword{javascript,}
