% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{issomethinginthismatrix}
\alias{issomethinginthismatrix}
\title{Is something in a matrix.}
\usage{
issomethinginthismatrix(x, searchfor)
}
\arguments{
\item{x}{The matrix that needs to be searched.}

\item{searchfor}{Items to be searched for in matrix x.}
}
\description{
This function searches to see if item 'searchfor' is within the matrix 'x' and returns a respective true (T) and false (F).
}
\examples{
issomethinginthismatrix(x = data.frame(matrix(1:9, nrow = 3, ncol = 3)), 
                       searchfor = 9)
issomethinginthismatrix(x = data.frame(matrix(LETTERS[1:9], nrow = 3, ncol = 3)), 
                       searchfor = "J")
}
\keyword{footnote,}
\keyword{footnotes}
\keyword{matrix,}
\keyword{search,}
