% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modnum}
\alias{modnum}
\title{Modify number.}
\usage{
modnum(x, divideby = 1000, commaseperator = T)
}
\arguments{
\item{x}{A numeric.}

\item{divideby}{Default = 1000.}

\item{commaseperator}{Default = T.}
}
\description{
Modify number.
}
\examples{
modnum()
}
\keyword{Modify}
\keyword{number}
