% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{modnum}
\alias{modnum}
\title{Modify number.}
\usage{
modnum(x, divideby = 1000, commaseperator = TRUE, digits = 0)
}
\arguments{
\item{x}{A numeric.}

\item{divideby}{The value you want all of your values divided by. Default = 1000.}

\item{commaseperator}{Do you want numbers to have commas in it ("1,000" (T) vs. "1000" (F). Default = T.}

\item{digits}{How many digits you would like your number to have. Default = 0.}
}
\description{
Modify number.
}
\examples{
modnum(x = data.frame(matrix(data = c(20000678660, 234567, 1, NA, 2345, 23), 
       ncol = 2)))
modnum(x = data.frame(matrix(data = c(20000678660, 234567, 1, NA, 2345, 23), 
                             ncol = 2)), 
       commaseperator = FALSE)
modnum(x = data.frame(matrix(data = c(200000, 234567, 1, NA, 2345, 23))), 
       divideby = 1, 
       digits = 2)
}
\keyword{Modify}
\keyword{number}
