% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{numbers0}
\alias{numbers0}
\title{Name nth item in order (001)}
\usage{
numbers0(x)
}
\arguments{
\item{x}{a numeric or integer value or vector of numeric or integer values.}
}
\value{
the values in the "0..X" format. All values will take on the number of 0s of the longest charcter value.
}
\description{
Convert a value.
}
\examples{
numbers0(x = c(1, 3, 6, 101))
}
\keyword{Data}
\keyword{Management}
