% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{numbers2words}
\alias{numbers2words}
\title{Convert number to text string.}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{The numbers that need to be converted to string.}
}
\description{
Convert number to text string.
}
\examples{
numbers2words()
}
\keyword{Text}
\keyword{editing}
