% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pchange}
\alias{pchange}
\title{Calculate the percent change.}
\usage{
pchange(start, end, ending = "", percent_first = TRUE)
}
\arguments{
\item{start}{The value it started with.}

\item{end}{The value it ended with.}

\item{ending}{A text string. Default "".}

\item{percent_first}{Options: T/F.}
}
\description{
Calculate the percent change.
}
\examples{
pchange()
}
\keyword{Modify}
\keyword{number}
