% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{save_graphs}
\alias{save_graphs}
\title{Systematically save your ggplot figure for your report}
\usage{
save_graphs(
  plot0,
  plot_list,
  header = "",
  footnote = "",
  filename0 = "x",
  cnt_chapt_content = "001",
  cnt = 1,
  path = NULL,
  width = 6,
  height = 6,
  output_type = c("pdf", "png"),
  type = "Figure",
  filename_desc = "",
  nickname = "",
  message = FALSE
)
}
\arguments{
\item{plot0}{The ggplot you would like to be saved}

\item{plot_list}{The list where all plots will be saved.}

\item{header}{The name and title of the figure. Default = "".}

\item{footnote}{Any footnote you want attached to this figure.}

\item{filename0}{The filename set at the begining of the chapter}

\item{cnt_chapt_content}{The order number that this exists in the chapter}

\item{cnt}{The figure number}

\item{path}{The path the file needs to be saved to. Default = "NULL", meaning it wont save anything and will override all other saving elements.}

\item{width}{Default = 6 inches}

\item{height}{Default = 6 inches}

\item{output_type}{Default = c("pdf", "png"). Can be anything supported by ggsave()}

\item{type}{Default = "Figure", but can be anything that the element needs to be called (e.g., "Graphic", "Fig.", "Graph") to fit in the phrase "Figure 1. This is my plot!"}

\item{filename_desc}{Additional description text for the filename that will be added at the name of file before the filename extention. Can be use to add a species name, location, or anything else that would make it easier to know what that file shows.}

\item{nickname}{A unique name that can be used to identify the figure so it can be referenced later in the report.}

\item{message}{TRUE/FALSE. Default = FALSE. If TRUE, it will print information about where your plot has been saved to.}
}
\value{
plot_list updated with the new plot and metadata.
}
\description{
Systematically save your ggplot figure for your report
}
\examples{
library(magrittr)
library(ggplot2)
plot_list <- c()
dat <- data.frame(x = rnorm(n = 10),
                  y = rnorm(n = 10),
                  col = rep_len(x = c("a", "b"),
                                length.out = 5))
# Select data and make plot
plot0<- dat \%>\%
  ggplot(aes(x = x, y = y,
  colour = as.factor(col))) + # create plot
  geom_point()
plot_list<-save_graphs(plot0 = plot0,
                      plot_list = plot_list,
                      header = "example",
                      footnote = "footnote example")
names(plot_list)
plot_list
}
