% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SaveGraphs}
\alias{SaveGraphs}
\title{Systematically save your ggplot figure for your report}
\usage{
SaveGraphs(
  plot0,
  plot.list,
  filename0,
  cnt.chapt.content,
  cnt.figures,
  path,
  width = 6,
  height = 6
)
}
\arguments{
\item{plot0}{The ggplot you would like to be saved}

\item{filename0}{The filename for your chapter}

\item{cnt.chapt.content}{The order number that this exists in the chapter}

\item{cnt.figures}{The figure number}

\item{width}{Default = 6 inches}

\item{height}{Default = 6 inches}
}
\value{

}
\description{
Systematically save your ggplot figure for your report
}
