% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SaveTables}
\alias{SaveTables}
\title{Systematically save your report tables for your report}
\usage{
SaveTables(
  table.raw = NULL,
  table.print,
  Header,
  Footnotes = NA,
  filename00,
  dir.chapters,
  dir.tables
)
}
\arguments{
\item{table.raw}{Optional. The data.frame that has no rounding and no dividing of numbers (good to save this for record keeping)}

\item{table.print}{The data.frame as table will be seen in the report.}

\item{Header}{The header or title of your table}

\item{dir.chapters}{Directory where you are saving all of your chapter word documents to.}

\item{dir.tables}{Directory where you are saving all of your tables to.}

\item{filename0}{The name you want to save this file as.}

\item{cnt.chapt.content}{The order number that this exists in the chapter}

\item{cnt.tables}{The figure table}
}
\value{

}
\description{
Systematically save your report tables for your report
}
