% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{SaveTables}
\alias{SaveTables}
\title{Systematically save your report tables for your report}
\usage{
SaveTables(
  table.raw = NULL,
  table.print,
  Header = "",
  Footnotes = NA,
  filename0 = "x",
  dir.chapters = NULL,
  dir.tables = NULL,
  cnt.chapt.content = "001"
)
}
\arguments{
\item{table.raw}{Optional. The data.frame that has no rounding and no dividing of numbers (good to save this for record keeping). Default = NA.}

\item{table.print}{The data.frame as table will be seen in the report.}

\item{Header}{The header or title of your table}

\item{Footnotes}{Footnotes for the whole table. Default = NA.}

\item{filename0}{The name you want to save this file as.}

\item{dir.chapters}{Directory where you are saving all of your chapter word documents to. Defult = NULL, meaning that it wont save anything.}

\item{dir.tables}{Directory where you are saving all of your tables to. Defult = NULL, meaning that it wont save anything.}

\item{cnt.chapt.content}{The order number that this exists in the chapter}
}
\value{
table.print
}
\description{
Systematically save your report tables for your report
}
\examples{
table.print<-data.frame(matrix(data = 1:9, nrow = 3, byrow = 3))
SaveTables(table.print=table.print)
}
