% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{SaveTables}
\alias{SaveTables}
\title{Systematically save your report tables for your report}
\usage{
SaveTables(
  table_raw = NULL,
  table_print,
  table_list = c(),
  Title = "",
  Footnotes = NA,
  filename0 = "x",
  dir_out_chapters = NULL,
  dir_tables = NULL,
  cnt_tables = 1,
  cnt_chapt_content = "001"
)
}
\arguments{
\item{table_raw}{Optional. The data.frame that has no rounding and no dividing of numbers (good to save this for record keeping). Default = NA.}

\item{table_print}{The data.frame as table will be seen in the report.}

\item{table_list}{Save tables in a list}

\item{Title}{The header or title of your table.}

\item{Footnotes}{Footnotes for the whole table. Default = NA.}

\item{filename0}{The name you want to save this file as.}

\item{dir_out_chapters}{Directory where you are saving all of your chapter word documents to. Defult = NULL, meaning that it wont save anything.}

\item{dir_tables}{Directory where you are saving all of your tables to. Defult = NULL, meaning that it wont save anything.}

\item{cnt_tables}{The order number that this exists in the chapter.}

\item{cnt_chapt_content}{The order number that this exists in the chapter.}
}
\description{
Systematically save your report tables for your report
}
\examples{
table_print<-data.frame(matrix(data = 1:9, nrow = 3, byrow = 3))
SaveTables(table_print=table_print)
}
