% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{text_list}
\alias{text_list}
\title{List items in sentance form.}
\usage{
text_list(x)
}
\arguments{
\item{x}{The text strings}
}
\description{
List items in sentance form.
}
\examples{
text_list()
}
\keyword{Text}
\keyword{editing}
