% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{TitleCase}
\alias{TitleCase}
\title{Make a String Title Case}
\usage{
TitleCase(str = "", add_dont_cap = "")
}
\arguments{
\item{str}{A string that you want to be in title case}

\item{add_dont_cap}{A vector of strings that the user does not want capitalized}
}
\description{
Make a String Title Case (making and, the, an, etc. lower case)
}
\examples{
TitleCase("HelLo WoRLD OR good-bye?")
}
\keyword{Case,}
\keyword{Title,}
\keyword{strings}
\keyword{word}
