% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFSReports.R
\name{tolower2}
\alias{tolower2}
\title{Make a string lower case except for stated proper nouns.}
\usage{
tolower2(str0, capitalizefirst = FALSE, add_cap = "")
}
\arguments{
\item{str0}{The text string.}

\item{capitalizefirst}{Default = FALSE}

\item{add_cap}{A vector of strings that the user does not want capitalized}
}
\description{
Make a string lower case except for stated proper nouns.
}
\examples{
tolower2(str0 = "notice how there are built-in proper nouns are capitalized: 
alaska is not in the south atlantic.", 
         capitalizefirst = TRUE,
         add_cap = "Proper nouns")
}
\keyword{Text}
\keyword{editing}
