% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xunits}
\alias{xunits}
\title{Determine the appropriate unit for a value.}
\usage{
xunits(temp00, combine = T)
}
\arguments{
\item{temp00}{A numeric.}

\item{combine}{Should this be combined in a single string (T) or as two seperate strings in a list (F). Default = T.}
}
\description{
Determine the appropriate unit for a value (e.g., 1000000 = '1 Million'.
}
\examples{
xunits()
}
\keyword{Modify}
\keyword{number,}
\keyword{units}
