% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xunitspct}
\alias{xunitspct}
\title{Determine the appropriate unit for a percent value.}
\usage{
xunitspct(value, sign = T)
}
\arguments{
\item{value}{A numeric.}

\item{sign}{Include percent sign. Default = T.}
}
\description{
Determine the appropriate unit for a percent value (e.g., 1000000 = '1 Million'.
}
\examples{
xunitspct()
}
\keyword{Modify}
\keyword{number,}
\keyword{units}
